+++ test/runtests.jl
@@ -8,4 +8,5 @@ using Test
     include("test_bbm_bbm_variable_bathymetry_1d.jl")
     include("test_svaerd_kalisch_1d.jl")
     include("test_serre_green_naghdi_1d.jl")
+    include("test_hyperbolic_serre_green_naghdi_1d.jl")
 end